requirements.txt
cloudinary==1.26.0
matplotlib==3.4.2
Flask==2.0.1
numpy==1.19.5
opencv-python-headless==4.5.2.54
python-dotenv==0.18.0
Pillow==8.0.1
Werkzeug==2.0.1
gunicorn
